Доделал добавление события. Осталось связать всё с бд

+++ eventclass.cpp
@@ -5,7 +5,7 @@ EventClass::EventClass()
 
 }
 
-EventClass::EventClass(int type, int points, int k, QDate date)
+EventClass::EventClass(int type, float points, int k, QDate date)
 {
     this->type = type;
     this->points = points;
